Cookies with firebaseauth
