+++ src/react/pages/HomeScreen.jsx
@@ -8,11 +8,16 @@ import ReorderSection from '../sections/ReorderSection';
 import Navbar from '../components/navbar/Navbar';
 import '@/sass/pages/_home_screen.scss';
 
+import { useLocation } from 'react-router-dom';
+import { showNotification } from '@mantine/notifications';
+
 const HomeScreen = () => {
   const [trendingLooks, setTrendingLooks] = useState([]);
   const [loadingTrending, setLoadingTrending] = useState(true);
   const [errorTrending, setErrorTrending] = useState(null);
+  const location = useLocation();
 
+  // 🔄 Fetch trending items
   useEffect(() => {
     fetch('/api/items?category_group=trending')
       .then(res => {
@@ -23,6 +28,17 @@ const HomeScreen = () => {
       .catch(err => setErrorTrending(err.message))
       .finally(() => setLoadingTrending(false));
   }, []);
+
+  // ✅ Show welcome message once after redirect
+  useEffect(() => {
+    if (location.state?.newUser) {
+      showNotification({
+        title: '🎉 Welcome!',
+        message: 'Your account was created successfully!',
+        color: 'green',
+      });
+    }
+  }, [location.state]);
 
   return (
     <div className='home-screen'>
