+++ src/react/pages/LoginPage.jsx
@@ -29,58 +29,40 @@ const LoginPage = () => {
   const navigate = useNavigate();
 
   // New helper: authenticate against backend
-  const loginWithBackend = async (email, password) => {
-    try {
-      const res = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ email, password }),
-      });
-      const data = await res.json();
-      if (!res.ok) throw new Error(data.error || 'Login failed');
-
-      localStorage.setItem('userId', data.user_id);
-      showNotification({
-        title: 'Welcome back!',
-        message: 'You are now logged in.',
-        color: 'green',
-      });
-      navigate('/homescreen');
-    } catch (err) {
-      console.error('❌ Backend login error:', err.message);
-      showNotification({
-        title: 'Login error',
-        message: err.message,
-        color: 'red',
-      });
-    }
-  };
-
+  
   const handleEmailLogin = async () => {
-    if (!email || !password) {
-      showNotification({
-        title: 'Missing credentials',
-        message: 'Please enter both email and password.',
-        color: 'red',
-      });
-      return;
-    }
+  if (!email || !password) {
+    showNotification({
+      title: 'Missing credentials',
+      message: 'Please enter both email and password.',
+      color: 'red',
+    });
+    return;
+  }
 
-    setLoading(true);
-    try {
-      const { auth, signInWithEmailAndPassword } = await getFirebaseAuth();
-      await signInWithEmailAndPassword(auth, email, password);
-      await loginWithBackend(email, password);
-    } catch (err) {
-      showNotification({
-        title: 'Login error',
-        message: err.message,
-        color: 'red',
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
+  setLoading(true);
+  try {
+    const { auth, signInWithEmailAndPassword } = await getFirebaseAuth();
+    await signInWithEmailAndPassword(auth, email, password);
+
+    showNotification({
+      title: 'Welcome back!',
+      message: 'You are now logged in.',
+      color: 'green',
+    });
+
+    navigate('/profile'); // or /homescreen
+  } catch (err) {
+    showNotification({
+      title: 'Login error',
+      message: err.message,
+      color: 'red',
+    });
+  } finally {
+    setLoading(false);
+  }
+};
+
 
   const handleGoogleSignIn = async () => {
     setLoading(true);
