+++ src/react/pages/signup/PersonalLookPage.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useLocation } from 'react-router-dom';
+import { showNotification } from '@mantine/notifications';
 import FinalStepper from '../../components/stepper/Stepper';
 import { BackIconButton } from '../../components/buttons/IconButtons';
 import { BottomBarButton } from '../../components/buttons/RedButtons';
